separated Book and User Controller
- updated main classes
- updated test classes
