+++ bookshop-project2/src/e2e/java/com/feri/alessandro/attsw/project/e2e/BookWebControllerE2E.java
@@ -10,7 +10,7 @@ import io.github.bonigarcia.wdm.WebDriverManager;
 
 @RunWith(Cucumber.class)
 @CucumberOptions(features = "src/e2e/resources/operations.feature", monochrome = true)
-public class OperationsE2E {
+public class BookWebControllerE2E {
 
 	@BeforeClass
 	public static void setUpDriver() {
