+++ bookshop-project2/src/it/java/com/feri/alessandro/attsw/project/BookWebControllerIT.java
@@ -4,7 +4,6 @@ import static java.util.Arrays.asList;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.hamcrest.CoreMatchers.nullValue;
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
 import static org.springframework.security.test.web.servlet.setup.SecurityMockMvcConfigurers.springSecurity;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.model;
@@ -22,32 +21,22 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.context.annotation.Import;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.context.junit4.SpringRunner;
-import org.springframework.test.web.ModelAndViewAssert;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.web.context.WebApplicationContext;
 
 import com.feri.alessandro.attsw.project.model.Book;
-import com.feri.alessandro.attsw.project.model.User;
 import com.feri.alessandro.attsw.project.repositories.BookRepository;
-import com.feri.alessandro.attsw.project.repositories.UserRepository;
 import com.feri.alessandro.attsw.project.security.SecurityConfiguration;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
 @Import(SecurityConfiguration.class)
-public class BookshopWebControllerIT {
+public class BookWebControllerIT {
 	
 	private static final String EMPTY_MESSAGE = "";
-
-	@Autowired
-	private UserRepository userRepository;
-	
-	@Autowired
-	private BCryptPasswordEncoder bCryptPasswordEncoder;
 	
 	@Autowired
 	private BookRepository bookRepository;
@@ -65,93 +54,13 @@ public class BookshopWebControllerIT {
 				webAppContextSetup(context).
 					apply(springSecurity()).
 						build();
-		
-		userRepository.deleteAll();
+
 		bookRepository.deleteAll();
 	}
 	
 	@After
 	public void tearDown() {
-		userRepository.deleteAll();
 		bookRepository.deleteAll();
-	}
-	
-	
-	@Test
-	public void test_returnLoginPageView() throws Exception {		
-		ModelAndViewAssert.assertViewName(
-				mvc.perform(get("/login")).
-					andReturn().getModelAndView(), "login");
-	}
-	
-	@Test
-	public void test_returnRegistrationPageView() throws Exception {
-		ModelAndViewAssert.assertViewName(
-				mvc.perform(get("/registration")).
-					andReturn().getModelAndView(), "registration");
-	}
-	
-	@Test
-	public void test_createNewUser_shouldReturnResultPage() throws Exception {
-		ModelAndViewAssert.assertViewName(mvc.perform(post("/saveUser")
-				.param("email", "email@gmail")
-				.param("username", "test")
-				.param("password", "password")).andReturn().getModelAndView(), "registrationResult");
-		
-		assertEquals(1, userRepository.findAll().size());
-		
-		User saved = userRepository.findAll().get(0);
-		
-		assertEquals("email@gmail", saved.getEmail());
-		assertEquals("test", saved.getUsername());
-		assertTrue(bCryptPasswordEncoder.matches("password", saved.getPassword()));
-		
-	}
-	
-	@Test
-	public void test_createNewUser_WhenEmailAlreadyExistShouldNotSaveTheUser_andShouldReturnResultPage() throws Exception {
-		User saved = new User();
-		saved.setEmail("already_exist@gmail");
-		saved.setUsername("username");
-		saved.setPassword("password");
-		
-		userRepository.save(saved);
-		
-		ModelAndViewAssert.assertViewName(mvc.perform(post("/saveUser")
-				.param("email", "already_exist@gmail")
-				.param("username", "user")
-				.param("password", "pass")).andReturn().getModelAndView(), "registrationResult");
-		
-		assertEquals(1, userRepository.findAll().size());
-		assertThat(userRepository.findAll()).containsExactly(saved);
-		
-		
-	}
-	
-	@Test
-	public void test_createNewUser_WhenUsernamaAlreadyExistShouldNotSaveTheUser_andShouldReturnResultPage() throws Exception {
-		User saved = new User();
-		saved.setEmail("email@gmail");
-		saved.setUsername("already_exist");
-		saved.setPassword("password");
-		
-		userRepository.save(saved);
-		
-		ModelAndViewAssert.assertViewName(mvc.perform(post("/saveUser")
-				.param("email", "not_exist@gmail")
-				.param("username", "already_exist")
-				.param("password", "pass")).andReturn().getModelAndView(), "registrationResult");
-		
-		assertEquals(1, userRepository.findAll().size());
-		assertThat(userRepository.findAll()).containsExactly(saved);
-	}
-	
-	@Test
-	@WithMockUser
-	public void test_returnHomePageView() throws Exception {
-		ModelAndViewAssert.assertViewName(
-				mvc.perform(get("/")).
-					andReturn().getModelAndView(), "index");
 	}
 	
 	@Test
@@ -206,7 +115,7 @@ public class BookshopWebControllerIT {
 			andExpect(model().attribute("book", nullValue())).
 			andExpect(model().attribute("message", "Book not found!"));
 		
-		assertThat(userRepository.findAll()).isEmpty();
+		assertThat(bookRepository.findAll()).isEmpty();
 	}
 	
 	@Test
