+++ bookshop-project2/src/main/java/com/feri/alessandro/attsw/project/web/BookWebController.java
@@ -10,89 +10,52 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.servlet.ModelAndView;
 
 import com.feri.alessandro.attsw.project.exception.BookNotFoundException;
-import com.feri.alessandro.attsw.project.exception.EmailExistException;
-import com.feri.alessandro.attsw.project.exception.UsernameExistException;
 import com.feri.alessandro.attsw.project.model.Book;
 import com.feri.alessandro.attsw.project.model.BookDTO;
-import com.feri.alessandro.attsw.project.model.User;
-import com.feri.alessandro.attsw.project.model.UserDTO;
 import com.feri.alessandro.attsw.project.services.BookService;
-import com.feri.alessandro.attsw.project.services.UserService;
-
 
 @Controller
-public class BookshopWebController {
-	
+public class BookWebController {
+
 	private static final String REDIRECT = "redirect:/";
 
 	private static final String MESSAGE = "message";
 
 	private static final String EMPTY_MESSAGE = "";
-
+	
 	@Autowired
 	private BookService bookService;
 	
-	@Autowired
-	private UserService userService;
-	
-	@GetMapping("/login")
-	public ModelAndView getLoginPage() {
-		return new ModelAndView("login");
-	}
-	
-	@GetMapping("/registration")
-	public ModelAndView getRegistrationPage() {
-		return new ModelAndView("registration");
-	}
-	
-	@PostMapping("/saveUser")
-	public ModelAndView createNewUser(UserDTO userDTO, Model model) throws EmailExistException, UsernameExistException {
-		User user = new User();
-		user.setEmail(userDTO.getEmail());
-		user.setUsername(userDTO.getUsername());
-		user.setPassword(userDTO.getPassword());
-		
-		userService.findUserByEmail(user.getEmail());
-		userService.findUserByUsername(user.getUsername());
-		
-		model.addAttribute(MESSAGE, EMPTY_MESSAGE);
-		userService.saveUser(user);
-		
-		return new ModelAndView("registrationResult");
-	}
-	
-	
 	@GetMapping("/")
-	public ModelAndView getIndex(Model model) {
+	public String getIndex(Model model) {
 		List<Book> allBooks = bookService.getAllBooks();
 		model.addAttribute("books", allBooks);
 		model.addAttribute(MESSAGE, allBooks.isEmpty() ? "There are no books." : EMPTY_MESSAGE);
 		
-		return new ModelAndView("index");
+		return "index";
 	}
 
 	@GetMapping("/edit/{id}")
-	public ModelAndView editBookById(@PathVariable BigInteger id, Model model) throws BookNotFoundException {
+	public String editBookById(@PathVariable BigInteger id, Model model) throws BookNotFoundException {
 		Book book = bookService.getBookById(id);
 		model.addAttribute("book", book);
 		model.addAttribute(MESSAGE, EMPTY_MESSAGE);
 			
-		return new ModelAndView("edit");
+		return "edit";
 	}
 	
 	@GetMapping("/new")
-	public ModelAndView newBook(Model model) {
+	public String newBook(Model model) {
 		model.addAttribute("book", new Book());
 		model.addAttribute(MESSAGE, EMPTY_MESSAGE);
 		
-		return new ModelAndView("edit");
+		return "edit";
 	}
 	
 	@PostMapping("/save")
-	public ModelAndView saveBook(BookDTO bookDTO) throws BookNotFoundException {
+	public String saveBook(BookDTO bookDTO) throws BookNotFoundException {
 		Book book = new Book();
 		book.setId(bookDTO.getId());
 		book.setTitle(bookDTO.getTitle());
@@ -106,11 +69,11 @@ public class BookshopWebController {
 			bookService.editBookById(id, book);
 		}
 		
-		return new ModelAndView(REDIRECT);
+		return REDIRECT;
 	}
 	
 	@GetMapping("/search")
-	public ModelAndView search (@RequestParam("title_searched") String title, Model model) throws BookNotFoundException {
+	public String search (@RequestParam("title_searched") String title, Model model) throws BookNotFoundException {
 		if(title.equals("")) {
 			model.addAttribute(MESSAGE, "Error! Please, insert a valid title.");
 		} else {
@@ -121,22 +84,23 @@ public class BookshopWebController {
 		
 		}
 		
-		return new ModelAndView("search");
+		return "search";
 	}
 	
 	@GetMapping("/delete")
-	public ModelAndView deleteBook(@RequestParam("id") BigInteger id, Model model) throws BookNotFoundException {
+	public String deleteBook(@RequestParam("id") BigInteger id, Model model) throws BookNotFoundException {
 		Book toDelete = bookService.getBookById(id);
 		bookService.deleteOneBook(toDelete);
 		
-		return new ModelAndView(REDIRECT);
+		return REDIRECT;
 	}
 	
 	@GetMapping("/deleteAll")
-	public ModelAndView deleteAll() {
+	public String deleteAll() {
 		bookService.deleteAllBooks();
 	
-		return new ModelAndView(REDIRECT);
+		return REDIRECT;
 		
 	}
+
 }
